refactor: split MagicMissionView into smaller View Models
